[Decomp][Spell] Cohesion split: extract SpellChecks handlers to SpellChecks.cpp
